#38: Command line moved.
