+++ src/main/java/org/cqfn/patternika/util/cmdline/CmdLine.java
@@ -1,4 +1,4 @@
-package org.cqfn.patternika.launcher;
+package org.cqfn.patternika.util.cmdline;
 
 import java.util.List;
 import java.util.Map;
